fix(db): fanout/seq VAL is DBF_LONG — accept a caput, don't refuse it as an enum choice
fanout and seq are "no-output-value / trigger" records: VAL is
`field(VAL,DBF_LONG){ pp(TRUE) }` (fanoutRecord.dbd:21-25,
seqRecord.dbd:21-25). C `process` (fanoutRecord.c:92-158, seqRecord.c)
never reads or writes VAL; it only drives the forward links. So a bare
`record(fanout,"X"){}` + `caput X.VAL 1` stores 1 verbatim, pp(TRUE)
reprocesses (a no-op for a link-less record), and STAT/SEVR stay
NO_ALARM.

The port declared VAL as `#[field(type = "Enum")] u16` on both records.
An `Enum` field serves DBR_ENUM and, on a `DBR_STRING` put, routes the
string→native conversion through the enum choice-matcher — a bare "1"
is refused as `S_db_badChoice` ("Channel write request failed") and VAL
stays 0. Declaring VAL `DBF_LONG` (i32) routes the string put through
the numeric `c_parse::put_string` Long row, which stores the value and
range-checks it exactly as C's `epicsParseInt32` does. VAL is already in
each record's pp(TRUE) table (process_passive.rs), so acceptance is the
only thing that was missing; forward-link processing was already wired.

Regression: tests/fanout_seq_val_long_put_stores.rs proves caput VAL
{zero, one, negative-one, i32::MIN, i32::MAX} stores the value and leaves
a bare fanout AND seq record at STAT/SEVR NO_ALARM.

+++ crates/epics-base-rs/src/server/records/seq.rs
@@ -28,8 +28,16 @@ const SEQ_SELM_CHOICES: &[&str] = &["All", "Specified", "Mask"];
     init = seq_init_record
 )]
 pub struct SeqRecord {
-    #[field(type = "Enum")]
-    pub val: u16,
+    // VAL is `field(VAL,DBF_LONG){ pp(TRUE) }` (seqRecord.dbd:21-25) — the "Used
+    // to trigger" field. It carries no output value: C `process`
+    // (seqRecord.c) never reads or writes VAL, it only sequences the DOn→LNKn
+    // writes. A `caput X.VAL 1` therefore stores 1 verbatim and pp(TRUE)
+    // reprocesses. Declaring it `DBF_LONG` (i32) routes a `DBR_STRING` put
+    // through the numeric `c_parse::put_string` Long row; the previous
+    // `Enum`/u16 declaration sent it to the enum choice-matcher, which refused a
+    // bare "1" as S_db_badChoice.
+    #[field(type = "Long")]
+    pub val: i32,
     // OLDN is `DBF_USHORT` (seqRecord.dbd:54), change-detection state for the
     // SELN monitor: C `process` posts a SELN monitor when `seln != oldn`, then
     // copies `oldn = seln` (seqRecord.c:230-232). Its init value is `seln`
